shared: extract code from lib into utils
